pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "export_wandb_to_mlflow"
authors = [{ name = "Chen Qian", email = "chen.qian@databricks.com" }]
requires-python = ">=3.8"
version = "0.0.1.dev0"
description = "Tools to migrate wandb projects to mlflow."
readme = "README.rst"
dependencies = ["mlflow-skinny>=2.9.0", "wandb>=0.16.0", "absl-py==2.1.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["export_wandb_to_mlflow", "export_wandb_to_mlflow.*"]
exclude = ["tests", "tests.*"]

[project.scripts]
w2m = "export_wandb_to_mlflow.main:main"

[tool.ruff]
line-length = 100

[tool.isort]
line_length = 100
